Etat du mercredi matin avant finalisation des non-reg de python
